requirements.txt
streamlit
Flask
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
deep_translator==1.10.1